feat: Update .gitignore file with additional rules
This commit adds a more comprehensive set of rules to the .gitignore file. The new rules cover various operating systems and development tools, including Linux, macOS, Node.js, Windows, and Yarn. The changes were made using the '.gitignore Generator' for Visual Studio Code and the templates provided by Toptal. For a detailed list of the added rules, please refer to the diff in the commit.